# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "taiyaq-backend",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/orders/display": {
      "get": {
        "tags": ["display"],
        "summary": "GET /api/orders/display",
        "operationId": "get_display_orders",
        "responses": {
          "200": {
            "description": "Orders grouped for display",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayOrdersResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to load display orders"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["display"],
        "summary": "GET /api/orders/{id}",
        "operationId": "get_order_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Failed to load order details"
          }
        }
      }
    },
    "/api/staff/flavors/config": {
      "get": {
        "tags": ["staff"],
        "summary": "GET /api/staff/flavors/config",
        "operationId": "get_flavor_configs",
        "responses": {
          "200": {
            "description": "Flavor configs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlavorConfigsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to load flavor configs"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/flavors/{flavor}": {
      "put": {
        "tags": ["staff"],
        "summary": "PUT /api/staff/flavors/{flavor}",
        "operationId": "set_flavor_config",
        "parameters": [
          {
            "name": "flavor",
            "in": "path",
            "description": "Flavor",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Flavor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlavorConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flavor config updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to save flavor config"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/orders": {
      "get": {
        "tags": ["staff"],
        "summary": "GET /api/staff/orders",
        "operationId": "get_staff_orders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Comma-separated order statuses",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff order list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffOrderResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to load orders"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["staff"],
        "summary": "POST /api/staff/orders",
        "operationId": "create_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrderResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to save order"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/orders/{id}/cancel": {
      "post": {
        "tags": ["staff"],
        "summary": "POST /api/staff/orders/{id}/cancel",
        "operationId": "cancel_order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrderResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Failed to save cancelled order"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/orders/{id}/complete": {
      "post": {
        "tags": ["staff"],
        "summary": "POST /api/staff/orders/{id}/complete",
        "operationId": "complete_order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completed order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrderResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Failed to save completed order"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/orders/{id}/notification": {
      "put": {
        "tags": ["staff"],
        "summary": "PUT /api/orders/{id}/notification",
        "operationId": "add_notification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated order notification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrderResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Failed to save notification update"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/orders/{id}/priority": {
      "put": {
        "tags": ["staff"],
        "summary": "PUT /api/staff/orders/{id}/priority",
        "operationId": "update_order_priority",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderPriorityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffOrderResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Failed to save order priority update"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/production": {
      "post": {
        "tags": ["staff"],
        "summary": "POST /api/staff/production",
        "operationId": "update_production",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Production update result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProductionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to save production update"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/staff/stock": {
      "get": {
        "tags": ["staff"],
        "summary": "GET /api/staff/stock",
        "operationId": "get_stock",
        "responses": {
          "200": {
            "description": "Unallocated stock by flavor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to load stock"
          }
        },
        "security": [
          {
            "staffBearerAuth": []
          }
        ]
      }
    },
    "/api/wait-times": {
      "get": {
        "tags": ["display"],
        "summary": "GET /api/wait-times",
        "operationId": "get_wait_times",
        "responses": {
          "200": {
            "description": "Current wait times",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitTimeResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to load wait times"
          }
        }
      }
    },
    "/line_callback": {
      "post": {
        "tags": ["line"],
        "operationId": "line_callback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Webhook accepted"
          },
          "400": {
            "description": "Invalid webhook body"
          },
          "401": {
            "description": "Invalid LINE signature"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateOrderRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "isPriority": {
            "type": ["boolean", "null"]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "DisplayOrder": {
        "type": "object",
        "required": ["id", "displayNumber"],
        "properties": {
          "displayNumber": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "DisplayOrdersResponse": {
        "type": "object",
        "required": ["ready", "cooking", "waiting"],
        "properties": {
          "cooking": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisplayOrder"
            }
          },
          "ready": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisplayOrder"
            }
          },
          "waiting": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisplayOrder"
            }
          }
        }
      },
      "Flavor": {
        "type": "string",
        "enum": ["tsubuan", "custard", "kurikinton"]
      },
      "FlavorConfig": {
        "type": "object",
        "required": ["cookingTimeMinutes", "quantityPerBatch"],
        "properties": {
          "cookingTimeMinutes": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "quantityPerBatch": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "FlavorConfigsResponse": {
        "type": "object",
        "required": ["tsubuan", "custard", "kurikinton"],
        "properties": {
          "custard": {
            "$ref": "#/components/schemas/FlavorConfig"
          },
          "kurikinton": {
            "$ref": "#/components/schemas/FlavorConfig"
          },
          "tsubuan": {
            "$ref": "#/components/schemas/FlavorConfig"
          }
        }
      },
      "Item": {
        "type": "object",
        "required": ["flavor", "quantity"],
        "properties": {
          "flavor": {
            "$ref": "#/components/schemas/Flavor"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Notify": {
        "oneOf": [
          {
            "type": "object",
            "required": ["Discord"],
            "properties": {
              "Discord": {
                "type": "object",
                "required": ["channel_id", "user_id"],
                "properties": {
                  "channel_id": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  },
                  "user_id": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["Line"],
            "properties": {
              "Line": {
                "type": "object",
                "required": ["user_id"],
                "properties": {
                  "user_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
      "NotifyRequest": {
        "oneOf": [
          {
            "type": "object",
            "required": ["Discord"],
            "properties": {
              "Discord": {
                "type": "object",
                "required": ["channel_id", "user_id"],
                "properties": {
                  "channel_id": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  },
                  "user_id": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": ["Line"],
            "properties": {
              "Line": {
                "type": "object",
                "required": ["user_id"],
                "properties": {
                  "user_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
      "OrderDetailsResponse": {
        "type": "object",
        "required": ["id", "displayNumber", "items", "status", "orderedAt"],
        "properties": {
          "displayNumber": {
            "type": "string"
          },
          "estimatedWaitMinutes": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "orderedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "enum": ["waiting", "cooking", "ready", "completed", "cancelled"]
      },
      "StaffOrderResponse": {
        "type": "object",
        "required": [
          "id",
          "displayNumber",
          "items",
          "status",
          "orderedAt",
          "isPriority"
        ],
        "properties": {
          "completedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "displayNumber": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "isPriority": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "orderedAt": {
            "type": "string",
            "format": "date-time"
          },
          "readyAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        }
      },
      "StockResponse": {
        "type": "object",
        "required": ["tsubuan", "custard", "kurikinton"],
        "properties": {
          "custard": {
            "type": "integer",
            "minimum": 0
          },
          "kurikinton": {
            "type": "integer",
            "minimum": 0
          },
          "tsubuan": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "UpdateOrderPriorityRequest": {
        "type": "object",
        "required": ["isPriority"],
        "properties": {
          "isPriority": {
            "type": "boolean"
          }
        }
      },
      "UpdateProductionRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "UpdateProductionResponse": {
        "type": "object",
        "required": ["newlyReadyOrders", "unallocatedItems"],
        "properties": {
          "newlyReadyOrders": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          "unallocatedItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "WaitTimeResponse": {
        "type": "object",
        "required": ["waitTimes"],
        "properties": {
          "waitTimes": {
            "$ref": "#/components/schemas/WaitTimes"
          }
        }
      },
      "WaitTimes": {
        "type": "object",
        "properties": {
          "custard": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "kurikinton": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "tsubuan": {
            "type": ["integer", "null"],
            "format": "int64"
          }
        }
      }
    },
    "securitySchemes": {
      "staffBearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "tags": [
    {
      "name": "display",
      "description": "Public order display APIs"
    },
    {
      "name": "staff",
      "description": "Staff operation APIs"
    },
    {
      "name": "line",
      "description": "LINE webhook APIs"
    }
  ]
}